db_engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "db_engine"
version = "0.1.0"
description = "OLTP to OLAP Database Engine with WAL, MVCC, B+Tree indexing, and vectorized execution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Systems Engineering Portfolio"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database :: Database Engines/Servers",
]
dependencies = [
    "pydantic>=2.0",
    "structlog>=24.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    "prometheus-client>=0.19",
    "numpy>=1.26",
    "numba>=0.59",
    "sqlglot>=25.0",
    "grpcio>=1.60",
    "xxhash>=3.4",
    "sortedcontainers>=2.4",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.98",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.2",
    "pre-commit>=3.6",
]

[project.scripts]
db-engine = "db_engine.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["db_engine"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (with infrastructure)",
    "chaos: Chaos/fault injection tests",
    "property: Property-based tests",
    "security: Security tests",
    "benchmark: Performance benchmarks",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src/db_engine"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
